[#24] refactor: ImageUseCase 분리
